requirements.txt
cffi==1.2.1
cryptography==1.0.1
Django==1.8.4
django-filter==0.11.0
djangorestframework==3.2.3
enum34==1.0.4
idna==2.0
ipaddress==1.0.14
ndg-httpsclient==0.4.0
psycopg2==2.6.1
pyasn1==0.1.8
pycparser==2.14
pyOpenSSL==0.15.1
six==1.9.0
wheel==0.24.0
